project/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
discord.py==2.3.1
exceptiongroup==1.1.2
frozenlist==1.4.0
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
requests==2.28.1
tomli==2.0.1
urllib3==1.26.12
yarl==1.9.2
